refactor: remove ModalMod mod file and move it to index
