+++ src/common/pure/Modal/index.tsx
@@ -8,8 +8,106 @@ import styled, { css } from 'styled-components/macro'
 
 import { isMobile } from 'legacy/utils/userAgent'
 
+export const HeaderRow = styled.div`
+  ${({ theme }) => theme.flexRowNoWrap};
+  padding: 1rem 1rem;
+  font-weight: 500;
+  color: ${(props) => (props.color === 'blue' ? ({ theme }) => theme.primary1 : 'inherit')};
+  ${({ theme }) => theme.mediaWidth.upToMedium`
+    padding: 1rem;
+  `};
+`
+
+export const CloseIcon = styled.div`
+  position: absolute;
+  right: 1rem;
+  top: 14px;
+  &:hover {
+    cursor: pointer;
+    opacity: 0.6;
+  }
+`
+
+export const HoverText = styled.div`
+  text-decoration: none;
+  color: ${({ theme }) => theme.text1};
+  display: flex;
+  align-items: center;
+
+  :hover {
+    cursor: pointer;
+  }
+`
+
+export const ContentWrapper = styled.div`
+  /* background-color: ${({ theme }) => theme.bg0}; */
+  background-color: ${({ theme }) => theme.bg1};
+  padding: 0 1rem 1rem 1rem;
+  border-bottom-left-radius: 20px;
+  border-bottom-right-radius: 20px;
+  ${({ theme }) => theme.mediaWidth.upToMedium`padding: 0 1rem 1rem 1rem`};
+`
+
+export const GpModal = styled(Modal)<{
+  maxWidth?: number | string
+  backgroundColor?: string
+  border?: string
+  padding?: string
+}>`
+  > [data-reach-dialog-content] {
+    color: ${({ theme }) => theme.text1};
+    width: 100%;
+    max-width: ${({ maxWidth = 470 }) => `${maxWidth}px`};
+    border: ${({ border = 'inherit' }) => `${border}`};
+    z-index: 100;
+    padding: ${({ padding = '0px' }) => `${padding}`};
+    margin: auto;
+    transition: max-width 0.4s ease;
+    background-color: ${({ theme }) => theme.bg1};
+    overflow: hidden;
+
+    ${({ theme }) => theme.mediaWidth.upToSmall`
+      max-height: 100vh;
+      max-width: 100%;
+      height: 100%;
+      width: 100vw;
+      border-radius: 0;
+    `}
+
+    ${HeaderRow} {
+      ${({ theme }) => theme.mediaWidth.upToSmall`
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        padding: 16px;
+        background: ${({ theme }) => theme.bg1};
+        z-index: 20;
+      `}
+    }
+
+    ${CloseIcon} {
+      ${({ theme }) => theme.mediaWidth.upToSmall`
+        z-index: 21;
+        position: fixed;
+      `}
+    }
+
+    ${HoverText} {
+      ${({ theme }) => theme.mediaWidth.upToSmall`
+        white-space: nowrap;
+      `}
+    }
+
+    ${ContentWrapper} {
+      ${({ theme }) => theme.mediaWidth.upToSmall`
+        margin: 62px auto 0;
+      `}
+    }
+  }
+`
+
 const AnimatedDialogOverlay = animated(DialogOverlay)
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 const StyledDialogOverlay = styled(AnimatedDialogOverlay)`
   &[data-reach-dialog-overlay] {
     z-index: 2;
@@ -86,7 +184,7 @@ interface ModalProps {
   children?: React.ReactNode
 }
 
-export default function Modal({
+export function Modal({
   isOpen,
   onDismiss,
   minHeight = false,
