+++ src/legacy/components/TokenWarningModal/index.tsx
@@ -2,7 +2,7 @@ import { Token } from '@uniswap/sdk-core'
 
 import { ImportToken } from 'legacy/components/SearchModal/ImportToken'
 
-import Modal from 'common/pure/Modal'
+import { Modal } from 'common/pure/Modal'
 
 export default function TokenWarningModal({
   isOpen,
